Fixed typos and small bugs

+++ utils.py
@@ -120,7 +120,7 @@ def colorEncode(labelmap, colors, mode='RGB'):
             continue
         labelmap_rgb += (labelmap == label)[:, :, np.newaxis] * \
             np.tile(colors[label+1],
-                    (labelmap.shape[0], labelmap.shape[1], 1))
+                    (labelmap.shape[0], labelmap.shape[1], 1)).astype(np.uint8)
 
     if mode == 'BGR':
         return labelmap_rgb[:, :, ::-1]
